live stacking
This uses the new nina live stacking plugin.

It adds a new variable live_stack to enable or disable live stacking. This is checked at the start of each target.

Before Target Scheduler starts a target it will check the variable and start live stacking if enabled.

After each target it will stop live stacking.

This is currently experimental.